Split `run.rs` into modules
